Seperated the Streamer class from oandapy in streamer/streamer.py For backwards compatibility __init__.py includes the Streamer class
