+++ oandapy/__init__.py
@@ -1,1 +1,2 @@
-from oandapy import *
+from oandapy import API, OandaError
+from stream.stream import Streamer
